Change precision to single decimal
Also updates plugin to allow changing temp ranges to one that
encroaches the minimum delta, by shifting the OTHER temperature
bound.

Finally, implements a loader script that prevents caching of JS
by adding a timestamp query param. This is different from the
last attempt of this because it ensures that scripts are loaded
in a deterministic order, which can be used to prevent dependency
issues. Dependencies can now also be moved out of the main html
file, and into separate script files; this will be deferred for
now, since we don't support adding event handlers to the load
event for scripts.